requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyMuPDF==1.23.8
tiktoken==0.5.1
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
ctransformers==0.2.27
Werkzeug==2.3.7 